Add cross-platform wheels and secure releases

+++ python/fasttext_module/fasttext/pybind/fasttext_pybind.cc
@@ -169,7 +169,7 @@ PYBIND11_MODULE(fasttext_pybind, m) {
       py::call_guard<py::gil_scoped_release>());
 
   py::class_<fasttext::Vector>(m, "Vector", py::buffer_protocol())
-      .def(py::init<ssize_t>())
+      .def(py::init<int64_t>())
       .def_buffer([](fasttext::Vector& m) -> py::buffer_info {
         return py::buffer_info(
             m.data(),
@@ -183,7 +183,7 @@ PYBIND11_MODULE(fasttext_pybind, m) {
   py::class_<fasttext::DenseMatrix>(
       m, "DenseMatrix", py::buffer_protocol(), py::module_local())
       .def(py::init<>())
-      .def(py::init<ssize_t, ssize_t>())
+      .def(py::init<int64_t, int64_t>())
       .def_buffer([](fasttext::DenseMatrix& m) -> py::buffer_info {
         return py::buffer_info(
             m.data(),
